+++ pluginlib/src/class_loader_base.cpp
@@ -1,4 +1,4 @@
-// Copyright 2008, Willow Garage, Inc. All rights reserved.
+// Copyright 2012, Willow Garage, Inc. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions are met:
@@ -26,45 +26,11 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-#ifndef PLUGINLIB__CLASS_DESC_HPP_
-#define PLUGINLIB__CLASS_DESC_HPP_
-
-#include <string>
-
-#include "pluginlib/visibility_control.hpp"
+#include "pluginlib/class_loader_base.hpp"
 
 namespace pluginlib
 {
 
-/// Storage for information about a given class.
-class PLUGINLIB_PUBLIC ClassDesc
-{
-public:
-  /**
-   * \param lookup_name The lookup name of the class
-   * \param derived_class The type of the derived class of the class
-   * \param base_class The type of the class, corresponds to the type of the base class
-   * \param package The package the class lives in
-   * \param description A description for the class
-   * \param library_name The name of the containing library for the class (not a full path!)
-   * \param plugin_manifest_path The path to the plugin manifest file
-   */
-  ClassDesc(
-    const std::string & lookup_name, const std::string & derived_class,
-    const std::string & base_class, const std::string & package,
-    const std::string & description, const std::string & library_name,
-    const std::string & plugin_manifest_path);
-
-  std::string lookup_name_;
-  std::string derived_class_;
-  std::string base_class_;
-  std::string package_;
-  std::string description_;
-  std::string library_name_;
-  std::string resolved_library_path_;  // This is set by pluginlib::ClassLoader at load time.
-  std::string plugin_manifest_path_;
-};
+ClassLoaderBase::~ClassLoaderBase() = default;
 
 }  // namespace pluginlib
-
-#endif  // PLUGINLIB__CLASS_DESC_HPP_
